Refactor local account directory. Add proxy account in db
