+++ core-db/src/main/java/io/novafoundation/nova/core_db/model/ExternalBalanceLocal.kt
@@ -3,7 +3,7 @@ package io.novafoundation.nova.core_db.model
 import androidx.room.Entity
 import androidx.room.ForeignKey
 import io.novafoundation.nova.core_db.model.chain.ChainAssetLocal
-import io.novafoundation.nova.core_db.model.chain.MetaAccountLocal
+import io.novafoundation.nova.core_db.model.chain.account.MetaAccountLocal
 import java.math.BigInteger
 
 @Entity(
