+++ core-db/src/main/java/io/novafoundation/nova/core_db/model/StakingDashboardItemLocal.kt
@@ -5,7 +5,7 @@ import androidx.room.Entity
 import androidx.room.ForeignKey
 import io.novafoundation.nova.core_db.model.chain.ChainAssetLocal
 import io.novafoundation.nova.core_db.model.chain.ChainLocal
-import io.novafoundation.nova.core_db.model.chain.MetaAccountLocal
+import io.novafoundation.nova.core_db.model.chain.account.MetaAccountLocal
 import jp.co.soramitsu.fearless_utils.runtime.AccountId
 import java.math.BigInteger
