+++ core-db/src/main/java/io/novafoundation/nova/core_db/model/WalletConnectPairingLocal.kt
@@ -4,7 +4,7 @@ import androidx.room.ColumnInfo
 import androidx.room.Entity
 import androidx.room.ForeignKey
 import androidx.room.PrimaryKey
-import io.novafoundation.nova.core_db.model.chain.MetaAccountLocal
+import io.novafoundation.nova.core_db.model.chain.account.MetaAccountLocal
 
 @Entity(
     tableName = "wallet_connect_pairings",
