+++ core-db/src/main/java/io/novafoundation/nova/core_db/model/chain/account/JoinedMetaAccountInfo.kt
@@ -1,0 +1,19 @@
+package io.novafoundation.nova.core_db.model.chain.account
+
+import androidx.room.Embedded
+import androidx.room.Relation
+
+interface JoinedMetaAccountInfo {
+
+    val metaAccount: MetaAccountLocal
+
+    val chainAccounts: List<ChainAccountLocal>
+}
+
+class RelationJoinedMetaAccountInfo(
+    @Embedded
+    override val metaAccount: MetaAccountLocal,
+
+    @Relation(parentColumn = "id", entityColumn = "metaId", entity = ChainAccountLocal::class)
+    override val chainAccounts: List<ChainAccountLocal>,
+) : JoinedMetaAccountInfo
