+++ feature-account-impl/src/main/java/io/novafoundation/nova/feature_account_impl/data/repository/ParitySignerRepository.kt
@@ -2,7 +2,7 @@ package io.novafoundation.nova.feature_account_impl.data.repository
 
 import io.novafoundation.nova.core.model.CryptoType
 import io.novafoundation.nova.core_db.dao.MetaAccountDao
-import io.novafoundation.nova.core_db.model.chain.MetaAccountLocal
+import io.novafoundation.nova.core_db.model.chain.account.MetaAccountLocal
 import io.novafoundation.nova.feature_account_api.domain.model.PolkadotVaultVariant
 import jp.co.soramitsu.fearless_utils.runtime.AccountId
