+++ feature-account-impl/src/main/java/io/novafoundation/nova/feature_account_impl/data/repository/WatchOnlyRepository.kt
@@ -1,8 +1,8 @@
 package io.novafoundation.nova.feature_account_impl.data.repository
 
 import io.novafoundation.nova.core_db.dao.MetaAccountDao
-import io.novafoundation.nova.core_db.model.chain.ChainAccountLocal
-import io.novafoundation.nova.core_db.model.chain.MetaAccountLocal
+import io.novafoundation.nova.core_db.model.chain.account.ChainAccountLocal
+import io.novafoundation.nova.core_db.model.chain.account.MetaAccountLocal
 import io.novafoundation.nova.runtime.multiNetwork.chain.model.ChainId
 import jp.co.soramitsu.fearless_utils.runtime.AccountId
